start rewriting full pipeline
