+++ src/test/scala/apps/cameraPipeCheck.scala
@@ -6,19 +6,13 @@ import rise.core._
 import rise.core.types._
 import rise.core.DSL._
 import rise.core.TypeLevelDSL._
-import rise.core.primitives._
 import elevate.core._
 import elevate.core.strategies.basic._
 import elevate.core.strategies.traversal._
-import elevate.core.strategies.predicate
 import elevate.rise.Rise
 import elevate.rise.rules._
 import elevate.rise.strategies.normalForm._
 import elevate.rise.rules.traversal._
-// import elevate.rise.strategies.traversal._
-import elevate.rise.rules.movement._
-import elevate.rise.rules.algorithmic._
-// import elevate.rise.strategies.predicate._
 
 class cameraPipeCheck extends shine.test_util.TestsWithExecutor {
   val H = 99
@@ -156,16 +150,6 @@ int main(int argc, char** argv) {
     printTime("execute", util.Execute(testCode))
   }
 
-  def dotPrintTmp(name: String, e: Rise): Unit = {
-    val generateDot = (e: Rise) => {
-      rise.core.dotPrinter.generateDotString(e,
-        printTypes = false,
-        inlineLambdaIdentifier = true,
-        applyNodes = false)
-    }
-    rise.core.dotPrinter.exprToDot("/tmp", name, e, generateDot)
-  }
-
   test("hot pixel suppression passes checks") {
     val typed = printTime("infer", infer(hot_pixel_suppression))
     println(s"hot pixel suppression: ${typed.t}")
@@ -236,471 +220,9 @@ int main(int argc, char** argv) {
   }
 
   test("demosaic passes checks with circular buffers") {
-    val typed = printTime("infer", infer(demosaic))
-
-    var nRewrite = 0
-    def rewrite(e: Rise, s: Strategy[Rise]): Rise = {
-      nRewrite += 1
-      val r = printTime(s"rewrite $nRewrite", s(e).get)
-      // dotPrintTmp(s"demosaic$nRewrite", r)
-      r
-    }
-
-    // 1. normalize a bit
-    val demosaic1 = rewrite(typed, normalize.apply(gentleBetaReduction))
-
-    def gentleFmap(s: Strategy[Rise]): Strategy[Rise] =
-      mapFusion `;` function(argument(body(s)) `;` mapLastFission)
-
-    // 2. push take/drop towards input
-    val demosaic2 = rewrite(demosaic1, normalize.apply(
-      gentleBetaReduction <+ etaReduction <+
-      takeAll <+ dropNothing <+ mapIdentity <+
-      takeBeforeMap <+ dropBeforeMap <+
-      gentleFmap(takeBeforeMap <+ dropBeforeMap) <+
-      takeInZip <+ dropInZip <+
-      takeInSelect <+ dropInSelect <+
-      (mapFusion `;`
-       function(argument(body(
-         normalize.apply(gentleBetaReduction) `;`
-         (takeInZip <+ dropInZip)
-       ))) `;`
-       fBeforeZipMap
-      ) <+ mapFBeforeSlide
+    checkDemosaic(cameraPipeRewrite.demosaicCircularBuffers(
+      printTime("infer", infer(cameraPipe.demosaic))
     ))
-
-    // 3. push line mapping towards output
-    val demosaic3 = rewrite(demosaic2, body(body(body(
-      function(body(function(body(
-        repeat(oncebu(
-          gentleBetaReduction <+ etaReduction <+ mapFusion <+ mapOutsideZip
-        )) `;`
-        // generate/select 1.1
-        oncetd(function(predicate.isEqualTo(generate)) `;`
-          oncetd(one(function(predicate.isEqualTo(generate)) `;`
-            argument(body({ x =>
-              var exprFound: Rise = null
-              function(argument(argument(argument({ expr =>
-                exprFound = expr
-                Success(expr)
-              })))).apply(x).flatMapSuccess(
-                argument(argument(
-                  argument(argument(slideAfter2) `;` dropBeforeMap) `;` takeBeforeMap `;`
-                  argument(zipSndAfter(exprFound))
-                ) `;` mapFusion `;` mapFusion)
-              )
-            })) `;`
-            mapOutsideGenerateSelect
-          ))
-        ) `;`
-        // generate/select 1.2
-        oncetd(function(predicate.isEqualTo(generate)) `;`
-          oncetd(one(function(predicate.isEqualTo(generate)) `;`
-            argument(body({ x =>
-              var rightExpr: Rise = null
-              argument(argument(argument(function(argument({ expr =>
-                rightExpr = expr
-                Success(expr)
-              }))))).apply(x).flatMapSuccess({ x =>
-                var leftExpr: Rise = null
-                function(argument(argument(
-                  zipSame `;` argument(
-                  zipSwap `;` argument(
-                  zipRotate `;` argument(
-                    function(argument(zipSame)) `;`
-                    argument(
-                      function(argument({ expr =>
-                        leftExpr = expr
-                        zipSndAfter(rightExpr)(expr)
-                      })) `;`
-                      argument(mapIdentityAfter) `;` mapOutsideZip
-                    ) `;` mapOutsideZip
-                  ) `;` mapFusion) `;` mapFusion) `;` mapFusion
-                ) `;` mapFusion)).apply(x).flatMapSuccess(
-                argument(argument(
-                  argument(
-                    function(argument(zipFstAfter(leftExpr))) `;`
-                    argument(mapIdentityAfter) `;` mapOutsideZip
-                  ) `;`
-                  function(argument(mapIdentityAfter)) `;` mapOutsideZip
-                ) `;` mapFusion)
-                )
-              })
-            })) `;` mapOutsideGenerateSelect
-          ))
-        ) `;`
-        normalize.apply(gentleBetaReduction <+ etaReduction <+ removeTransposePair <+ mapFusion) `;`
-        // generate/select 1
-        oncetd(function(predicate.isEqualTo(generate)) `;`
-          argument(body({ x =>
-            var rightExpr: Rise = null
-            argument(argument(argument(function(argument({ expr =>
-              rightExpr = expr
-              Success(expr)
-            }))))).apply(x).flatMapSuccess({ x =>
-              var leftExpr: Rise = null
-              function(argument(
-                argument(
-                  function(argument(
-                    takeBeforeDrop `;` argument(takeInSlide) `;` dropBeforeMap
-                  )) `;`
-                  argument(argument(takeInSlide)) `;`
-                  argument(function(argument({ expr =>
-                    leftExpr = expr
-                    zipSndAfter(rightExpr)(expr)
-                  })) `;` mapOutsideZip) `;` mapOutsideZip
-                ) `;` mapFusion
-              )).apply(x).flatMapSuccess(
-                argument(argument(
-                  argument(
-                    function(argument(zipFstAfter(leftExpr))) `;`
-                    argument(mapIdentityAfter) `;` mapOutsideZip
-                  ) `;`
-                  function(argument(mapIdentityAfter)) `;` mapOutsideZip
-                ) `;` mapFusion)
-              )
-            })
-          })) `;` mapOutsideGenerateSelect
-        ) `;`
-        normalize.apply(gentleBetaReduction <+ etaReduction <+ removeTransposePair <+ mapFusion) `;`
-        // generate/select 2.1
-        oncetd(function(predicate.isEqualTo(generate)) `;`
-          oncetd(one(function(predicate.isEqualTo(generate)) `;`
-            argument(body({ x =>
-              var rightExpr: Rise = null
-              argument(argument({ expr =>
-                rightExpr = expr
-                Success(expr)
-              })).apply(x).flatMapSuccess({ x =>
-                var leftExpr: Rise = null
-                function(argument(argument({ expr =>
-                  leftExpr = expr
-                  zipSndAfter(rightExpr)(expr)
-                }) `;` mapFusion)).apply(x).flatMapSuccess(
-                  argument(argument(
-                    zipFstAfter(leftExpr)
-                  ) `;` mapFusion)
-                )
-              })
-            })) `;` mapOutsideGenerateSelect
-          ))
-        ) `;`
-        // generate/select 2.2
-        oncetd(function(predicate.isEqualTo(generate)) `;`
-          oncetd(one(function(predicate.isEqualTo(generate)) `;`
-            argument(body({ x =>
-              var rightExpr: Rise = null
-              argument(argument({ expr =>
-                rightExpr = expr
-                Success(expr)
-              })).apply(x).flatMapSuccess({ x =>
-                var leftExpr: Rise = null
-                function(argument(argument({ expr =>
-                  leftExpr = expr
-                  zipSndAfter(rightExpr)(expr)
-                }) `;` mapFusion)).apply(x).flatMapSuccess(
-                  argument(argument(
-                    zipFstAfter(leftExpr)
-                  ) `;` mapFusion)
-                )
-              })
-            })) `;` mapOutsideGenerateSelect
-          ))
-        ) `;`
-        normalize.apply(gentleBetaReduction <+ etaReduction <+ removeTransposePair <+ mapFusion) `;`
-        // generate/select 2
-        oncetd(function(predicate.isEqualTo(generate)) `;`
-          argument(body({ x =>
-            var rightExpr: Rise = null
-            argument(argument({ expr =>
-              rightExpr = expr
-              Success(expr)
-            })).apply(x).flatMapSuccess({ x =>
-              var leftExpr: Rise = null
-              function(argument(argument({ expr =>
-                leftExpr = expr
-                zipSndAfter(rightExpr)(expr)
-              }) `;` mapFusion)).apply(x).flatMapSuccess(
-                argument(argument(
-                  zipFstAfter(leftExpr)
-                ) `;` mapFusion)
-              )
-            })
-          })) `;` mapOutsideGenerateSelect
-        ) `;`
-        normalize.apply(gentleBetaReduction <+ etaReduction <+ removeTransposePair <+ mapFusion) `;`
-        // generate/select 3.1
-        oncetd(function(predicate.isEqualTo(generate)) `;`
-          oncetd(one(function(predicate.isEqualTo(generate)) `;`
-            argument(body({ x =>
-              var rightExpr: Rise = null
-              argument(argument(
-                zipSame `;` argument(
-                zipSwap `;` argument(
-                zipRotate `;` argument(
-                  function(argument(zipSame)) `;`
-                  argument(function(argument({ expr =>
-                    rightExpr = expr
-                    Success(expr)
-                  }))) `;`
-                  argument(mapIdentityAfter) `;` mapOutsideZip
-                ) `;` mapFusion) `;` mapFusion) `;` mapFusion
-              ) `;` mapFusion).apply(x).flatMapSuccess({ x =>
-                var leftExpr: Rise = null
-                function(argument(argument(
-                  argument(
-                    function(argument({ expr =>
-                      leftExpr = expr
-                      zipSndAfter(rightExpr)(expr)
-                    })) `;`
-                    argument(mapIdentityAfter) `;` mapOutsideZip
-                  ) `;`
-                  function(argument(mapIdentityAfter)) `;` mapOutsideZip
-                ) `;` mapFusion)).apply(x).flatMapSuccess(
-                  argument(argument(
-                    argument(
-                      function(argument(zipFstAfter(leftExpr))) `;`
-                      argument(mapIdentityAfter) `;` mapOutsideZip
-                    ) `;`
-                    function(argument(mapIdentityAfter)) `;` mapOutsideZip
-                  ) `;` mapFusion)
-                )
-              })
-            })) `;` mapOutsideGenerateSelect
-          ))
-        ) `;`
-        // generate/select 3.2
-        oncetd(function(predicate.isEqualTo(generate)) `;`
-          oncetd(one(function(predicate.isEqualTo(generate)) `;`
-            argument(body({ x =>
-              var rightExpr: Rise = null
-              argument(argument(argument({ expr =>
-                rightExpr = expr
-                Success(expr)
-              }))).apply(x).flatMapSuccess(
-              function(argument(argument(
-                takeBeforeDrop `;`
-                argument(argument(slideAfter2) `;` takeBeforeMap) `;` dropBeforeMap `;`
-                argument(zipSndAfter(rightExpr)) `;` mapFusion
-              ) `;` mapFusion))
-              )
-            })) `;` mapOutsideGenerateSelect
-          ))
-        ) `;`
-        normalize.apply(gentleBetaReduction <+ etaReduction <+ removeTransposePair <+ mapFusion) `;`
-        // generate/select 3
-        oncetd(function(predicate.isEqualTo(generate)) `;`
-          argument(body({ x =>
-            var leftExpr: Rise = null
-            function(argument(argument(
-              argument(function(argument({ expr =>
-                leftExpr = expr
-                Success(expr)
-              })))
-            ))).apply(x).flatMapSuccess({ x =>
-              var rightExpr: Rise = null
-              argument(argument(
-                function(argument(dropBeforeTake `;` argument(dropInSlide) `;` takeBeforeMap)) `;`
-                argument(
-                  argument(dropInSlide) `;`
-                  function(argument({ expr =>
-                    rightExpr = expr
-                    zipFstAfter(leftExpr)(expr)
-                  })) `;`
-                  mapOutsideZip
-                ) `;` mapOutsideZip
-              ) `;` mapFusion).apply(x).flatMapSuccess(
-                function(argument(argument(
-                  argument(
-                    function(argument(zipSndAfter(rightExpr))) `;`
-                    argument(mapIdentityAfter) `;` mapOutsideZip
-                  ) `;`
-                  function(argument(mapIdentityAfter)) `;` mapOutsideZip
-                ) `;` mapFusion))
-              )
-              })
-          })) `;` mapOutsideGenerateSelect
-        ) `;`
-        normalize.apply(gentleBetaReduction <+ etaReduction <+ removeTransposePair <+ mapFusion) `;`
-        // makeArray
-        // makeArray .1
-        { x =>
-          var expr1: Rise = null
-          function(function(argument(argument(argument(
-            // zip ordering
-            argument(zipRotate) `;`
-            zipSwap `;`
-            argument(argument(mapIdentityAfter) `;` mapOutsideZip `;` argument(zipRotate)) `;`
-            mapFusion `;` mapFusion `;`
-            argument(
-              argument(
-                zipSwap `;` argument(
-                  argument(
-                    function(argument(zipSwap)) `;`
-                    argument(mapIdentityAfter) `;` mapOutsideZip `;` argument(zipRotate) `;`
-                    mapFusion `;`
-                    argument(argument(zipSwap)) `;`
-                    argument(function(argument(mapIdentityAfter)) `;` mapOutsideZip) `;`
-                    mapFusion
-                  ) `;`
-                  function(argument(mapIdentityAfter)) `;` mapOutsideZip
-                ) `;` mapFusion
-              ) `;`
-              function(argument(mapIdentityAfter)) `;` mapOutsideZip
-            ) `;` mapFusion `;`
-            argument(
-              // zip branch unification
-              argument(
-                function(argument({ expr =>
-                  expr1 = expr
-                  Success(expr)
-                })) `;`
-                argument(
-                  argument(
-                    argument(
-                      argument(slideAfter2) `;` dropBeforeMap `;` argument(dropInSlide) `;` mapFusion
-                    ) `;` function(argument(mapIdentityAfter)) `;` mapOutsideZip
-                  ) `;` function(argument(mapIdentityAfter)) `;` mapOutsideZip
-                ) `;` function(argument(mapIdentityAfter)) `;` mapOutsideZip
-              ) `;` function(argument(mapIdentityAfter)) `;` mapOutsideZip
-            )
-          ) `;` repeat(mapFusion))))).apply(x).flatMapSuccess(
-          function(argument(argument(argument(
-            // zip ordering
-            zipRotateRight `;` argument(
-              argument(
-                argument(zipSwap) `;`
-                function(argument(mapIdentityAfter)) `;` mapOutsideZip `;`
-                argument(zipRotateLeft `;` argument(
-                  function(argument(zipSwap)) `;`
-                  argument(mapIdentityAfter) `;` mapOutsideZip `;`
-                  argument(zipRotateRight)
-                ))
-              ) `;`
-              argument(repeatNTimes(3, mapFusion)) `;`
-              function(argument(mapIdentityAfter)) `;` mapOutsideZip
-            ) `;` mapFusion `;`
-            argument(
-              // zip branch unification
-              argument(
-                argument(
-                  argument(
-                    argument(slideAfter2) `;` dropBeforeMap `;` argument(dropInSlide) `;` mapFusion
-                  ) `;`
-                  function(argument(
-                    argument(slideAfter2) `;` takeBeforeMap `;` argument(takeInSlide) `;` mapFusion
-                  )) `;` mapOutsideZip
-                ) `;` function(argument(mapIdentityAfter)) `;` mapOutsideZip
-              ) `;` function(argument(mapIdentityAfter)) `;` mapOutsideZip
-            )
-          ) `;` repeatNTimes(2, mapFusion)))) `;`
-          { x: Rise =>
-            var expr2: Rise = null
-            argument(argument(argument(
-              // zip ordering
-              argument(
-                function(argument(
-                  zipRotateRight `;` argument(argument(zipSwap))
-                )) `;`
-                argument(mapIdentityAfter) `;` mapOutsideZip `;`
-                argument(zipRotateRight `;` argument(argument(
-                  argument(mapIdentityAfter) `;` mapOutsideZip `;`
-                  argument(zipRotateRight)
-                )))
-              ) `;`
-              argument(mapFusion) `;` function(argument(mapIdentityAfter)) `;`
-              mapOutsideZip `;` argument(
-                zipRotateLeft `;` argument(
-                  function(argument(zipSwap))
-                ) `;`
-                argument(
-                  argument(mapFusion) `;` mapOutsideZip `;`
-                  argument(zipRotateRight)
-                )
-              )
-            ) `;` repeatNTimes(4, mapFusion) `;`
-            argument(
-              // zip branch unification
-              argument(
-                function(argument({ expr =>
-                  expr2 = expr
-                  Success(expr)
-                })) `;`
-                argument(
-                  argument(
-                    function(argument(
-                      argument(slideAfter2) `;` takeBeforeMap `;` argument(takeInSlide) `;` mapFusion
-                    )) `;` argument(mapIdentityAfter) `;` mapOutsideZip
-                  ) `;` function(argument(dropBeforeTake `;` mapIdentityAfter)) `;` mapOutsideZip
-                ) `;` function(argument(mapIdentityAfter)) `;` mapOutsideZip
-              ) `;` function(argument(mapIdentityAfter)) `;` mapOutsideZip
-            ) `;` mapFusion
-          )).apply(x).flatMapSuccess(
-          // zip branch injection
-          function(function(argument(argument(argument(
-            argument(
-              argument(zipFstAfter(expr2)) `;`
-              function(argument(mapIdentityAfter)) `;` mapOutsideZip
-            ) `;` function(argument(mapIdentityAfter)) `;` mapOutsideZip
-          ) `;` mapFusion)))) `;`
-          function(argument(argument(argument(
-            argument(
-              zipFstAfter(expr1) `;`
-              argument(
-                argument(zipFstAfter(expr2)) `;`
-                function(argument(mapIdentityAfter)) `;` mapOutsideZip
-              ) `;` mapFusion
-            ) `;` function(argument(mapIdentityAfter)) `;` mapOutsideZip
-          ) `;` mapFusion))) `;`
-          argument(argument(argument(
-            argument(zipFstAfter(expr1)) `;` function(argument(mapIdentityAfter)) `;` mapOutsideZip
-          ) `;` mapFusion))
-          )}
-          )
-        } `;`
-        fOutsideMakeArray `;` argument(
-          mapOutsideMakeArray `;` argument(function(argument(
-            normalize.apply(
-              betaReduction <+ etaReduction <+ removeTransposePair <+ mapFusion <+
-              idxReduction <+ fstReduction <+ sndReduction
-            )
-          )))
-        )
-      ))))
-    ))))
-
-    // 4. lowering with slideSeq
-    val demosaic4 = rewrite(demosaic3, body(body(body(
-      function(body(function(body(
-        argument(argument(
-          function(function(lowering.mapStream)) `;`
-          argument(argument(argument(argument(argument(repeatNTimes(2, oncetd(
-            lowering.slideSeq(SlideSeq.Indices, mapSeq(fun(x => x)))
-          ))))))) `;`
-          // TODO: use proper rewriting to achieve this
-          function(argument(body({ expr =>
-            Success(
-              expr |> transpose >> map(transpose) >>
-              // 2 bands of y. all x. rgb channels.
-              mapSeqUnroll(mapSeq(mapSeqUnroll(fun(x => x)))) >>
-              map(transpose) >> transpose
-            )
-          }))) `;`
-          normalize.apply(
-            betaReduction <+ etaReduction <+ removeTransposePair <+ mapFusion <+
-            idxReduction <+ fstReduction <+ sndReduction
-          )
-        ))
-      ))))
-    ))))
-
-    // TODO: check the generated code and if the loop nests are similar to:
-    // demosaic(h)(w) >> transpose >> map(transpose) >>
-    //   split(2) >> mapSeq(mapSeqUnroll(
-    //     mapSeq(
-    //       mapSeqUnroll(fun(x => x)))
-    // )) >> join >> map(transpose) >> transpose
-    checkDemosaic(demosaic4)
   }
 
   test("color correction passes checks") {
@@ -800,6 +322,13 @@ ${fName}(output, ${2*H}, ${2*W}, input, ${sharpen_strength});
       3 * 2*H * 2*W, "uint8_t", "sharpened.dump",
       0
     )
+  }
+
+  test("camera pipe passes checks with circular buffers") {
+    cameraPipeRewrite.circularBuffers(
+      printTime("infer", infer(cameraPipe.camera_pipe))
+    )
+    // TODO: check output
   }
 
   test("type inference") {
